few fixes(moved methods to model)
